CHG: rename nwb module
- all other load functions have no leading `_`

Changes to be committed:
	modified:   syncopy/io/__init__.py
	renamed:    syncopy/io/_load_nwb.py -> syncopy/load_nwb.py
