+++ syncopy/io/__init__.py
@@ -10,14 +10,14 @@ from . import (
     save_spy_container,
     load_ft,
     load_tdt,
-    _load_nwb
+    load_nwb
 )
 from .utils import *
 from .load_spy_container import *
 from .save_spy_container import *
 from .load_ft import *
 from .load_tdt import *
-from ._load_nwb import *
+from .load_nwb import *
 
 # Populate local __all__ namespace
 __all__ = []
@@ -25,4 +25,4 @@ __all__.extend(utils.__all__)
 __all__.extend(load_spy_container.__all__)
 __all__.extend(save_spy_container.__all__)
 __all__.extend(load_ft.__all__)
-__all__.extend(_load_nwb.__all__)
+__all__.extend(load_nwb.__all__)
